requirements.txt
streamlit
boto3
openai
PyPDF2
retrying
python-dotenv
